requirements.txt
Flask
pdfminer.six
docx2txt
scikit-learn